# docs/API.yaml
openapi: 3.0.3
info:
  title: Tasky API
  description: |
    REST API for Tasky — Mongolia's most trusted domestic service marketplace.

    ## Authentication
    All endpoints (unless noted) require a Bearer JWT in the `Authorization` header.
    Obtain tokens via the OTP login flow (`/auth/otp/*`).

    ## Pagination
    All list endpoints use **cursor-based pagination**.
    Pass `cursor` (opaque string) and `limit` (default 20, max 100) as query parameters.
    Responses wrap results in `{ data: [], cursor: { next, has_more } }`.

    ## Errors
    All errors return a standard JSON body: `{ code, message, trace_id }`.

    ## Currency
    All monetary amounts are integers in MNT (Mongolian Tugrik).

    ## Idempotency
    Critical state-changing endpoints require an `Idempotency-Key` header.
    Payment and wallet endpoints are documented as post-MVP deferred surfaces.
  version: 1.1.0
  contact:
    name: Tasky Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor for pagination. Pass the `next` value from a previous response.
      schema:
        type: string
...
  schemas:
...
    PresignedUrlResponse:
      type: object
      required: [upload_url, storage_key]
      properties:
        upload_url:
          type: string
          format: uri
          description: Presigned URL to upload the file directly to S3/MinIO.
        storage_key:
          type: string
          description: Storage key to reference this file in subsequent API calls.
          example: uploads/avatars/550e8400-e29b-41d4-a716-446655440000.jpg

    # ─── Identity ─────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /tasks/photos/upload-url:
    post:
      tags: [Tasks]
      summary: Get presigned URL for pre-create task photo upload
      description: |
        Returns a presigned URL to upload a task photo directly to S3/MinIO before task creation.
        Upload keys can be submitted in `photo_keys` during `POST /tasks`.
      operationId: getTaskPreCreatePhotoUploadUrl
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type]
              properties:
                content_type:
                  type: string
                  example: image/jpeg
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks:
    get:
      tags: [Tasks]
      summary: List open tasks
      description: |
        Returns a paginated list of OPEN tasks. Supports filtering by category and
        geospatial radius search. Location data is fuzzed for privacy.
      operationId: listTasks
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by category ID.
          schema:
            type: string
            format: uuid
        - name: lat
          in: query
          description: Latitude for radius search.
          schema:
            type: number
            format: double
        - name: lng
          in: query
          description: Longitude for radius search.
          schema:
            type: number
            format: double
        - name: radius_km
          in: query
          description: Search radius in kilometers (default 10, max 50).
          schema:
            type: number
            format: double
            default: 10
            maximum: 50
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of open tasks.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicTask'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tasks]
      summary: Create a new task
      description: |
        Creates a new task in OPEN status. Only authenticated customers can create tasks.
        The task must have a future scheduled_at date and a valid category.
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task details
      description: |
        Returns full task details. Exact location is only visible to the task owner
        and the booked tasker. Other users see approximate location only.
      operationId: getTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Task details.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Task'
                  - $ref: '#/components/schemas/PublicTask'
                description: |
                  Returns role-scoped task details.
                  - Owners and booked taskers receive full `Task` (exact location fields populated).
                  - Other authorized users receive redacted `PublicTask` (approximate location only).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update a task
      description: |
        Updates a task. Only the task owner can update, and only while the task is in OPEN status.
      operationId: updateTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task is not in OPEN status and cannot be updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}/cancel:
    post:
      tags: [Tasks]
      summary: Cancel a task
      description: |
        Cancels an OPEN task. Only the task owner can cancel.
        If the task has an active booking, use the booking cancellation endpoint instead.
      operationId: cancelTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Task cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task cannot be cancelled in its current status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}/photos/upload-url:
    post:
      tags: [Tasks]
      summary: Get presigned URL for task photo upload
      description: |
        Returns a presigned URL to upload a task photo directly to S3/MinIO.
        Maximum 3 photos per task.
      operationId: getTaskPhotoUploadUrl
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type]
              properties:
                content_type:
                  type: string
                  example: image/jpeg
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Applications ───────────────────────────────────────────
  /tasks/{id}/applications:
    post:
      tags: [Applications]
      summary: Apply to a task
      description: |
        Tasker applies to an OPEN task. Only verified taskers can apply.
        A conversation is automatically created between the customer and tasker.
      operationId: applyToTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 500
                  description: Cover message explaining why the tasker is a good fit.
      responses:
        '201':
          description: Application submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskApplication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not a verified tasker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Tasker already applied or task is not OPEN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Applications]
      summary: List applications for a task
      description: |
        Returns paginated list of applications for a task.
        Only the task owner can view all applications.
      operationId: listTaskApplications
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of applications.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskApplication'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /bookings:
    get:
      tags: [Bookings]
      summary: List my bookings
      description: |
        Returns paginated list of bookings for the authenticated user.
        Filter by role (as customer or tasker) and booking status.
      operationId: listBookings
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          description: View bookings as customer or tasker.
          schema:
            type: string
            enum: [customer, tasker]
        - name: status
          in: query
          description: Filter by booking status.
          schema:
            type: string
            enum: [ASSIGNED, COMPLETED, CANCELLED]
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of bookings.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...